+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Expose provider stats on API intrerface via `.stats`
+- Add Kusama 9220 upgrade block
 
 
 ## 8.6.2 May 30, 2022