+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Expose provider stats on API intrerface via `.stats`
+- Adjust flag extraction for latest phragmen types
 
 
 ## 8.6.2 May 30, 2022