+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes:
 
 - Expose provider stats on API intrerface via `.stats`
 - Adjust flag extraction for latest phragmen types
+- Add Kusama 9220 upgrade block
 
 
 ## 8.6.2 May 30, 2022