+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Expose provider stats on API intrerface via `.stats`
+- Adjust flag extraction for latest phragmen types
 - Add Kusama 9220 upgrade block