+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## master
+## 8.7.1 Jun 4, 2022
 
 Changes: